Extract more types for reuse
