+++ src/vs/workbench/api/common/extHostTypes/symbolInformation.ts
@@ -1,0 +1,87 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { URI } from '../../../../base/common/uri.js';
+import { es5ClassCompat } from './es5ClassCompat.js';
+import { Location } from './location.js';
+import { Range } from './range.js';
+
+export enum SymbolKind {
+	File = 0,
+	Module = 1,
+	Namespace = 2,
+	Package = 3,
+	Class = 4,
+	Method = 5,
+	Property = 6,
+	Field = 7,
+	Constructor = 8,
+	Enum = 9,
+	Interface = 10,
+	Function = 11,
+	Variable = 12,
+	Constant = 13,
+	String = 14,
+	Number = 15,
+	Boolean = 16,
+	Array = 17,
+	Object = 18,
+	Key = 19,
+	Null = 20,
+	EnumMember = 21,
+	Struct = 22,
+	Event = 23,
+	Operator = 24,
+	TypeParameter = 25
+}
+
+export enum SymbolTag {
+	Deprecated = 1
+}
+
+@es5ClassCompat
+export class SymbolInformation {
+
+	static validate(candidate: SymbolInformation): void {
+		if (!candidate.name) {
+			throw new Error('name must not be falsy');
+		}
+	}
+
+	name: string;
+	location!: Location;
+	kind: SymbolKind;
+	tags?: SymbolTag[];
+	containerName: string | undefined;
+
+	constructor(name: string, kind: SymbolKind, containerName: string | undefined, location: Location);
+	constructor(name: string, kind: SymbolKind, range: Range, uri?: URI, containerName?: string);
+	constructor(name: string, kind: SymbolKind, rangeOrContainer: string | undefined | Range, locationOrUri?: Location | URI, containerName?: string) {
+		this.name = name;
+		this.kind = kind;
+		this.containerName = containerName;
+
+		if (typeof rangeOrContainer === 'string') {
+			this.containerName = rangeOrContainer;
+		}
+
+		if (locationOrUri instanceof Location) {
+			this.location = locationOrUri;
+		} else if (rangeOrContainer instanceof Range) {
+			this.location = new Location(locationOrUri!, rangeOrContainer);
+		}
+
+		SymbolInformation.validate(this);
+	}
+
+	toJSON(): any {
+		return {
+			name: this.name,
+			kind: SymbolKind[this.kind],
+			location: this.location,
+			containerName: this.containerName
+		};
+	}
+}
