+++ src/vs/workbench/contrib/notebook/browser/view/cellParts/cellOutput.ts
@@ -26,7 +26,8 @@ import { CellContentPart } from '../cellPart.js';
 import { CodeCellRenderTemplate } from '../notebookRenderingCommon.js';
 import { CodeCellViewModel } from '../../viewModel/codeCellViewModel.js';
 import { NotebookTextModel } from '../../../common/model/notebookTextModel.js';
-import { CellUri, IOrderedMimeType, NotebookCellExecutionState, NotebookCellOutputsSplice, RENDERER_NOT_AVAILABLE, isTextStreamMime } from '../../../common/notebookCommon.js';
+import { CellUri, IOrderedMimeType, NotebookCellExecutionState, NotebookCellOutputsSplice, RENDERER_NOT_AVAILABLE } from '../../../common/notebookCommon.js';
+import { isTextStreamMime } from '../../../../../../base/common/mime.js';
 import { INotebookExecutionStateService } from '../../../common/notebookExecutionStateService.js';
 import { INotebookKernel } from '../../../common/notebookKernelService.js';
 import { INotebookService } from '../../../common/notebookService.js';
